+++ src/workers/parallelMeshBVH.worker.js
@@ -1,11 +1,11 @@
 import { MathUtils, BufferGeometry, BufferAttribute } from 'three';
-import { BYTES_PER_NODE } from '../../core/Constants.js';
-import { buildTree, generateIndirectBuffer } from '../../core/build/buildTree.js';
-import { countNodes, populateBuffer } from '../../core/build/buildUtils.js';
-import { computeTriangleBounds } from '../../core/build/computeBoundsUtils.js';
-import { getFullGeometryRange, getRootIndexRanges, getTriCount } from '../../core/build/geometryUtils.js';
-import { WorkerPool } from '../utils/WorkerPool.js';
-import { DEFAULT_OPTIONS } from '../../core/MeshBVH.js';
+import { BYTES_PER_NODE } from '../core/Constants.js';
+import { buildTree, generateIndirectBuffer } from '../core/build/buildTree.js';
+import { countNodes, populateBuffer } from '../core/build/buildUtils.js';
+import { computeTriangleBounds } from '../core/build/computeBoundsUtils.js';
+import { getFullGeometryRange, getRootIndexRanges, getTriCount } from '../core/build/geometryUtils.js';
+import { WorkerPool } from './utils/WorkerPool.js';
+import { DEFAULT_OPTIONS } from '../core/MeshBVH.js';
 
 let isRunning = false;
 let prevTime = 0;
